+++ app/Admin/Controllers/ProcessController.php
@@ -5,12 +5,11 @@ namespace App\Admin\Controllers;
 use App\Http\Controllers\Controller;
 use App\Models\ShopProduct;
 use App\Models\ShopProductDescription;
-use App\Models\ShopSpecialPrice;
 use Encore\Admin\Controllers\HasResourceActions;
 use Encore\Admin\Facades\Admin;
 use Encore\Admin\Layout\Content;
-use Excel;
 use Illuminate\Http\Request;
+use PhpOffice\PhpSpreadsheet\IOFactory;
 
 class ProcessController extends Controller
 {
@@ -22,6 +21,7 @@ class ProcessController extends Controller
      */
     public function index(Content $content)
     {
+
         return $content
             ->header('Index')
             ->description(' ')
@@ -30,64 +30,117 @@ class ProcessController extends Controller
 
     public function importProduct(Content $content, Request $request)
     {
+
         if ($request->getMethod() == 'POST') {
-            $validatedData = \Validator::make($request->all(), [
-                'import_file' => 'required|mimes:xlsx',
-            ]);
-            if ($validatedData->fails()) {
-                return redirect()->back()->withErrors($validatedData->errors());
-            } else {
-                Excel::load($request->file('import_file'), function ($reader) use ($request) {
-                    $reader->skipRows(1); //
-                    $arrError   = array();
-                    $arrSuccess = array();
-                    foreach ($reader->toArray() as $key => $row) {
-                        if (ShopProduct::where('sku', $row['sku'])->first()) {
-                            $arrError[] = $row['sku'] . ': already exist!';
-                        } else {
+            $case = $request->get('case');
+            switch ($case) {
+                case 'import_file_info':
+                    $reader        = IOFactory::createReader('Xls');
+                    $validatedData = \Validator::make($request->all(), [
+                        'import_file_info' => 'required|mimes:xlsx,xls',
+                    ]);
+                    if ($validatedData->fails()) {
+                        return redirect()->back()->withErrors($validatedData->errors());
+                    } else {
+
+                        $spreadsheet = $reader->load($request->file('import_file_info'));
+                        $sheet       = $spreadsheet->getActiveSheet();
+                        $maxCol      = $sheet->getHighestColumn();
+                        $maxRow      = $sheet->getHighestRow();
+                        $data        = $sheet->rangeToArray('A3:' . $maxCol . $maxRow, true, true, true, true);
+                        $arrError    = array();
+                        $arrSuccess  = array();
+                        // dd($data);
+                        foreach ($data as $key => $row) {
+                            $arrMapping                = array();
+                            $sku                       = $row['A'];
+                            $arrMapping['price']       = (int) $row['B'];
+                            $arrMapping['cost']        = (int) $row['E'];
+                            $arrMapping['stock']       = (int) $row['G'];
+                            $arrMapping['category_id'] = (int) $row['B'];
+                            $arrMapping['brand_id']    = (int) $row['C'];
+                            $arrMapping['vendor_id']   = (int) $row['D'];
                             try {
-                                $arrMapping                = array();
-                                $arrMapping['sku']         = $row['sku'];
-                                $arrMapping['price']       = (int) $row['price'];
-                                $arrMapping['cost']        = (int) $row['cost'];
-                                $arrMapping['stock']       = (int) $row['stock'];
-                                $arrMapping['category_id'] = (int) $row['category_id'];
-                                $arrMapping['brand_id']    = (int) $row['brand_id'];
-                                $arrMapping['vendor_id']   = (int) $row['vendor_id'];
-                                $id                        = ShopProduct::insertGetId($arrMapping);
-                                $descriptons               = [
-                                    'product_id'  => $id,
-                                    'name'        => $row['name'],
-                                    'description' => $row['description'],
-                                    'keyword'     => $row['keyword'],
-                                    'content'     => $row['content'],
-                                    'lang_id'     => (int) $row['language'],
-                                ];
-                                ShopProductDescription::insert($descriptons);
-                                if ($row['special_price']) {
-                                    ShopSpecialPrice::insert([
-                                        'product_id' => $id,
-                                        'price'      => $row['special_price'],
-                                        'status'     => 1,
-                                    ]);
-                                }
-                                $arrSuccess[] = $row['sku'];
+                                (new ShopProduct)
+                                    ->updateOrInsert(
+                                        ['sku' => $sku],
+                                        $arrMapping
+                                    );
+                                $arrSuccess[] = $sku;
                             } catch (\Exception $e) {
-                                // $arrError[] = $row['sku'] . ': have error ' . $e->getMessage();
-                                $arrError[] = $row['sku'] . ': have error';
+                                $arrError[] = $sku . ': ' . $e->getMessage();
                             }
                         }
+                        if ($arrSuccess) {
+                            $request->session()->flash('import_success', $arrSuccess);
+                        }
+                        if ($arrError) {
+                            $request->session()->flash('import_error', $arrError);
+                        }
+                        return back();
 
                     }
-                    if ($arrSuccess) {
-                        $request->session()->flash('import_success', $arrSuccess);
+
+                    break;
+
+                case 'import_file_description':
+                    $reader        = IOFactory::createReader('Xls');
+                    $validatedData = \Validator::make($request->all(), [
+                        'import_file_description' => 'required|mimes:xlsx,xls',
+                    ]);
+                    if ($validatedData->fails()) {
+                        return redirect()->back()->withErrors($validatedData->errors());
+                    } else {
+
+                        $spreadsheet = $reader->load($request->file('import_file_description'));
+                        $sheet       = $spreadsheet->getActiveSheet();
+                        $maxCol      = $sheet->getHighestColumn();
+                        $maxRow      = $sheet->getHighestRow();
+                        $data        = $sheet->rangeToArray('A3:' . $maxCol . $maxRow, true, true, true, true);
+                        $arrError    = array();
+                        $arrSuccess  = array();
+                        foreach ($data as $key => $row) {
+                            $sku          = $row['A'];
+                            $checkProduct = ShopProduct::where('sku', $sku)->first();
+                            if (!$checkProduct) {
+                                $arrError[] = $sku . ': already not exist!';
+                            } else {
+                                try {
+                                    $arrUnique   = ['product_id' => $checkProduct->id, 'lang_id' => (int) $row['B']];
+                                    $descriptons = [
+                                        'name'        => $row['C'],
+                                        'description' => $row['D'],
+                                        'keyword'     => $row['E'],
+                                        'content'     => $row['F'],
+                                    ];
+                                    (new ShopProductDescription)
+                                        ->updateOrInsert(
+                                            $arrUnique,
+                                            $descriptons
+                                        );
+                                    $arrSuccess[] = $sku;
+                                } catch (\Exception $e) {
+                                    // $arrError[] = $sku . ': have error ' . $e->getMessage();
+                                    $arrError[] = $sku . ': ' . $e->getMessage();
+                                }
+                            }
+
+                        }
+                        if ($arrSuccess) {
+                            $request->session()->flash('import_success', $arrSuccess);
+                        }
+                        if ($arrError) {
+                            $request->session()->flash('import_error', $arrError);
+                        }
+                        return back();
+
                     }
-                    if ($arrError) {
-                        $request->session()->flash('import_error', $arrError);
-                    }
-                    return back();
-                });
+                    break;
+                default:
+                    # code...
+                    break;
             }
+
         }
 
         return $content
@@ -98,7 +151,7 @@ class ProcessController extends Controller
 
     public function getImportProduct()
     {
-        return view('admin.ImportProduct')->render();
+        return view('admin.importProduct')->render();
 
     }