+++ resources/lang/vi/language.php
@@ -222,8 +222,13 @@ return [
     ],
 
     'process'            => [
-        'productImport'      => 'Thêm nhiều sản phẩm',
-        'productImport_text' => 'Tải file mẫu ở đây',
+        'productImport'            => 'Thêm mới/Cập nhật nhiều sản phẩm',
+        'product_info'             => 'Thông tin sản phẩm (Bước 1)',
+        'product_info_note'        => '<b>Ví dụ:</b> SKU, price, stock ...',
+        'product_description'      => 'Nội dung sản phẩm (Bước 2)',
+        'product_description_note' => '<b>Ví dụ:</b> Tên, từ khóa, mô tả...',
+        'productImport_text'       => 'Tải file mẫu',
+        'productImport_download'   => 'Tải Ở ĐÂY',
     ],
 
     'profile'            => [